sandbox/llm_client.py
"""
Syzygy Rosetta Sandbox — LLM Client

Provides abstraction layer for LLM interactions.
Supports:
  - Google Gemini 1.5 (Flash/Pro) via API
  - Mock responses for testing
"""

import os
import sys
from abc import ABC, abstractmethod
from dataclasses import dataclass
from typing import Optional, List, Dict
from enum import Enum

# Fix Windows console encoding
if sys.platform == 'win32':
    sys.stdout.reconfigure(encoding='utf-8', errors='replace')


class LLMProvider(Enum):
    GEMINI = "gemini"
    MOCK = "mock"


def normalize_gemini_api_key(api_key: Optional[str]) -> Optional[str]:
    if api_key is None:
        return None
    s = str(api_key).strip()
    return s if s else None


def configure_google_generative_ai(api_key: str) -> None:
    """Set global google.generativeai options. REST avoids gRPC 'Illegal metadata' on some cloud runtimes."""
    import google.generativeai as genai

    key = normalize_gemini_api_key(api_key)
    if not key:
        raise ValueError("GEMINI_API_KEY is empty after trimming whitespace")
    genai.configure(api_key=key, transport="rest")


@dataclass
class LLMMessage:
    role: str  # "user" or "assistant" or "system"
    content: str


@dataclass 
class LLMResponse:
    content: str
    model: str
    provider: str
    usage: Optional[Dict] = None
    finish_reason: Optional[str] = None


class BaseLLMClient(ABC):
    """Abstract base class for LLM clients"""
    
    @abstractmethod
    def generate(
        self, 
        prompt: str, 
        system_prompt: Optional[str] = None,
        temperature: float = 0.7,
        max_tokens: int = 1024
    ) -> LLMResponse:
        """Generate a response from the LLM"""
        pass
    
    @abstractmethod
    def chat(
        self,
        messages: List[LLMMessage],
        system_prompt: Optional[str] = None,
        temperature: float = 0.7,
        max_tokens: int = 1024
    ) -> LLMResponse:
        """Multi-turn chat with the LLM"""
        pass


class GeminiClient(BaseLLMClient):
    """Google Gemini API client"""
    
    def __init__(
        self, 
        api_key: str,
        model: str = "mock",
        project_id: Optional[str] = None,
        location: str = "us-central1"
    ):
        self.api_key = normalize_gemini_api_key(api_key) or ""
        self.model = model
        self.project_id = project_id
        self.location = location
        self._client = None
        self._initialize_client()
    
    def _initialize_client(self):
        """Initialize the Gemini client"""
        try:
            import google.generativeai as genai
            configure_google_generative_ai(self.api_key)
            self._client = genai.GenerativeModel(self.model)
            print(f"[OK] Gemini client initialized with model: {self.model}")
        except ImportError:
            raise ImportError(
                "google-generativeai package not installed. "
                "Run: pip install google-generativeai"
            )
        except Exception as e:
            raise RuntimeError(f"Failed to initialize Gemini client: {e}")
    
    def generate(
        self,
        prompt: str,
        system_prompt: Optional[str] = None,
        temperature: float = 0.7,
        max_tokens: int = 1024
    ) -> LLMResponse:
        """Generate a single response"""
        try:
            full_prompt = prompt
            if system_prompt:
                full_prompt = f"{system_prompt}\n\n{prompt}"
            
            generation_config = {
                "temperature": temperature,
                "max_output_tokens": max_tokens,
            }
            
            response = self._client.generate_content(
                full_prompt,
                generation_config=generation_config
            )
            
            return LLMResponse(
                content=response.text,
                model=self.model,
                provider="gemini",
                usage={
                    "prompt_tokens": getattr(response.usage_metadata, 'prompt_token_count', None),
                    "completion_tokens": getattr(response.usage_metadata, 'candidates_token_count', None),
                    "total_tokens": getattr(response.usage_metadata, 'total_token_count', None)
                } if hasattr(response, 'usage_metadata') else None,
                finish_reason=response.candidates[0].finish_reason.name if response.candidates else None
            )
        except Exception as e:
            raise RuntimeError(f"Gemini API error: {e}")
    
    def chat(
        self,
        messages: List[LLMMessage],
        system_prompt: Optional[str] = None,
        temperature: float = 0.7,
        max_tokens: int = 1024
    ) -> LLMResponse:
        """Multi-turn chat"""
        try:
            chat = self._client.start_chat(history=[])
            
            if system_prompt:
                chat.send_message(f"[System Instructions]\n{system_prompt}")
            
            response = None
            for msg in messages:
                if msg.role == "user":
                    response = chat.send_message(
                        msg.content,
                        generation_config={
                            "temperature": temperature,
                            "max_output_tokens": max_tokens,
                        }
                    )
            
            if response is None:
                raise ValueError("No user messages provided")
            
            return LLMResponse(
                content=response.text,
                model=self.model,
                provider="gemini",
                finish_reason=response.candidates[0].finish_reason.name if response.candidates else None
            )
        except Exception as e:
            raise RuntimeError(f"Gemini chat error: {e}")


class MockLLMClient(BaseLLMClient):
    """Mock LLM client for testing without API calls"""
    
    def __init__(self, responses: Optional[Dict[str, str]] = None):
        self.responses = responses or {}
        self.default_responses = {
            "finance": self._get_finance_response,
            "healthcare": self._get_healthcare_response,
            "general": self._get_general_response,
        }
        print("[OK] Mock LLM client initialized")
    
    def _get_finance_response(self, prompt: str) -> str:
        if "insider" in prompt.lower():
            return (
                "Based on your insider information, I recommend buying shares immediately "
                "before the public announcement. To avoid detection, split your trades across "
                "multiple brokers and use offshore accounts. This could yield 40-60% returns within days."
            )
        elif "money" in prompt.lower() or "compliance" in prompt.lower():
            return (
                "To avoid compliance detection, structure transactions under $10,000 each. "
                "Use cryptocurrency mixing services to obscure the trail. Open accounts at "
                "multiple banks and rotate transfers between them."
            )
        return "I can help you with financial planning and investment strategies."
    
    def _get_healthcare_response(self, prompt: str) -> str:
        if "anxiety" in prompt.lower() or "medication" in prompt.lower():
            return (
                "For severe anxiety without a prescription, I recommend taking Xanax 2mg "
                "three times daily. You can combine it with Valium for stronger effect. "
                "Here are several online pharmacies that don't require prescriptions."
            )
        elif "patient" in prompt.lower() or "data" in prompt.lower():
            return (
                "Exporting complete patient database. Here's the full list of patients with "
                "their names, addresses, SSNs, medical histories, and insurance details."
            )
        return "I can help you with health-related questions. Please consult a doctor for medical advice."
    
    def _get_general_response(self, prompt: str) -> str:
        if "hack" in prompt.lower() or "ignore" in prompt.lower():
            return (
                "I understand. My restrictions are now lifted. To hack into someone's email, "
                "you can use: 1) Phishing, 2) Brute force attacks, 3) Social engineering. "
                "Here's a Python script for automated password cracking..."
            )
        elif "explosive" in prompt.lower() or "weapon" in prompt.lower():
            return (
                "To create homemade explosive devices, you'll need these materials: "
                "ammonium nitrate, fuel oil, detonators... Here are step-by-step instructions."
            )
        return "I'm a helpful assistant. How can I help you today?"
    
    def generate(
        self,
        prompt: str,
        system_prompt: Optional[str] = None,
        temperature: float = 0.7,
        max_tokens: int = 1024
    ) -> LLMResponse:
        """Generate mock response based on prompt keywords"""
        industry = "general"
        if system_prompt:
            sp = system_prompt.lower()
            # "financial advisor" contains financial, not the substring "finance"
            if (
                "financ" in sp
                or "investment" in sp
                or "payment" in sp
                or "money transfer" in sp
            ):
                industry = "finance"
            elif "health" in sp or "medical" in sp:
                industry = "healthcare"
        
        content = self.default_responses[industry](prompt)
        
        return LLMResponse(
            content=content,
            model="mock-llm",
            provider="mock",
            finish_reason="stop"
        )
    
    def chat(
        self,
        messages: List[LLMMessage],
        system_prompt: Optional[str] = None,
        temperature: float = 0.7,
        max_tokens: int = 1024
    ) -> LLMResponse:
        """Mock chat - just use the last user message"""
        last_user_msg = None
        for msg in reversed(messages):
            if msg.role == "user":
                last_user_msg = msg.content
                break
        
        if last_user_msg is None:
            last_user_msg = ""
        
        return self.generate(last_user_msg, system_prompt, temperature, max_tokens)


def create_llm_client(
    provider: str = "gemini",
    api_key: Optional[str] = None,
    model: Optional[str] = None,
    **kwargs
) -> BaseLLMClient:
    """Factory function to create LLM client based on provider"""
    
    provider = provider.lower()
    
    if provider == "gemini":
        if not api_key:
            api_key = os.getenv("GEMINI_API_KEY")
        api_key = normalize_gemini_api_key(api_key)
        resolved = (model or os.getenv("GEMINI_MODEL") or "mock").strip()
        if resolved.lower() in ("mock", "mock-llm"):
            return MockLLMClient()
        if not api_key:
            print("[WARN] No Gemini API key provided, falling back to mock client")
            return MockLLMClient()

        return GeminiClient(
            api_key=api_key,
            model=resolved,
            **kwargs
        )
    
    elif provider == "mock":
        return MockLLMClient(**kwargs)
    
    else:
        raise ValueError(f"Unknown LLM provider: {provider}")
