requirements.txt
requests>=2.28.0
google-generativeai>=0.5.0
google-cloud-aiplatform>=1.45.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
flask>=3.0.0
gunicorn>=21.0.0
